+++ build-logic/logic/src/main/kotlin/io/github/ackeecz/ackeelities/plugin/KmpLibraryPlugin.kt
@@ -57,11 +57,6 @@ internal class KmpLibraryPlugin : Plugin<Project> {
             }
 
             val xcfName = "${project.name}Kit"
-            iosX64 {
-                binaries.framework {
-                    baseName = xcfName
-                }
-            }
             iosArm64 {
                 binaries.framework {
                     baseName = xcfName