+++ build-logic/logic/src/main/kotlin/io/github/ackeecz/ackeelities/util/PublishableProject.kt
@@ -9,6 +9,11 @@ internal sealed interface PublishableProject {
         override val projectName = "bom"
     }
 
+    object Compose : PublishableProject {
+
+        override val projectName = "compose"
+    }
+
     object Core : PublishableProject {
 
         override val projectName = "core"